refactor: rename test files and update namespaces to Refitter.Tests.Telemetry
